requirements.txt
pmdarima==2.0.4
numpy==1.26.4
matplotlib==3.7.1
statsmodels==0.14.0
scikit-learn==1.2.2
pandas==2.1.4
gspread==6.1.0
oauth2client==4.1.3
google-api-python-client==1.127.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
plotly==5.14.1
openpyxl==3.1.2
pyyaml==6.0.1
google-cloud-bigquery==3.21.0
mlflow==2.12.2